Fix tests, add do nothing in redistribution, iterating reductions, fee collection, tests
