+++ .env.example
@@ -14,7 +14,7 @@ SEPOLIA_URL =
 #Forkings
 MAINNET_FORKING_URL = 
 POLYGON_FORKING_URL = 
-
+OPTIMISM_FORKING_URL =
 #Other
 ETHERSCAN_API_KEY = 
 POLYGONSCAN_API_KEY =
