+++ hardhat.config.ts
@@ -1,14 +1,11 @@
 import * as dotenv from "dotenv";
 
 import { HardhatUserConfig } from "hardhat/config";
-import "@nomiclabs/hardhat-etherscan";
-import "@nomiclabs/hardhat-waffle";
-import "@typechain/hardhat";
-import "hardhat-gas-reporter";
-import "solidity-coverage";
 import 'hardhat-contract-sizer'
 import './tasks'
 import '@openzeppelin/hardhat-upgrades';
+
+import "@nomicfoundation/hardhat-toolbox";
 
 dotenv.config();
