+++ package.json
@@ -15,14 +15,11 @@
   },
   "homepage": "https://github.com/GetAlluo/yield-booster#readme",
   "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.2.1",
-    "@nomiclabs/hardhat-etherscan": "^3.1.0",
-    "@nomiclabs/hardhat-waffle": "^2.0.1",
+    "@nomicfoundation/hardhat-chai-matchers": "^1.0.6",
+    "@nomicfoundation/hardhat-toolbox": "^2.0.2",
     "@openzeppelin/contracts": "^4.4.1",
     "@openzeppelin/contracts-upgradeable": "^4.5.1",
     "@openzeppelin/hardhat-upgrades": "^1.21.0",
-    "@typechain/ethers-v5": "^7.2.0",
-    "@typechain/hardhat": "^2.3.1",
     "@types/chai": "^4.3.0",
     "@types/mocha": "^9.0.0",
     "@types/node": "^12.20.41",
@@ -30,7 +27,6 @@
     "@typescript-eslint/parser": "^4.33.0",
     "alluo-exchange": "github:GetAlluo/exchange",
     "alluo-strategies": "github:GetAlluo/investment-strategies",
-    "chai": "^4.3.4",
     "dotenv": "^10.0.0",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
@@ -39,19 +35,14 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^3.4.1",
     "eslint-plugin-promise": "^5.2.0",
-    "ethereum-waffle": "^3.4.0",
-    "ethers": "^5.7.2",
     "hardhat": "^2.12.2",
     "hardhat-contract-sizer": "^2.7.0",
-    "hardhat-gas-reporter": "^1.0.7",
     "husky": "^8.0.2",
     "prettier": "^2.8.0",
     "prettier-plugin-solidity": "^1.0.0",
     "solhint": "^3.3.7",
     "solhint-plugin-prettier": "^0.0.5",
-    "solidity-coverage": "^0.8.2",
     "ts-node": "^10.4.0",
-    "typechain": "^5.2.0",
     "typescript": "^4.5.4"
   },
   "dependencies": {
