+++ test/boost/AlluoVaultUpgradeableDolaFrax.ts
@@ -3,7 +3,7 @@ import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { expect } from "chai";
 import { ethers, network, upgrades } from "hardhat";
 import { afterEach, before } from "mocha";
-import { AlluoVaultUpgradeable, Exchange, AlluoVaultPool, IAlluoPool, ICurvePool, IERC20MetadataUpgradeable } from "../typechain";
+import { AlluoVaultUpgradeable, Exchange, AlluoVaultPool, IAlluoPool, ICurvePool, ICurveMetaPool, ICvxBooster, IERC20MetadataUpgradeable, IExchange } from "../../typechain";
 
 
 async function skipDays(d: number) {
@@ -11,17 +11,19 @@ async function skipDays(d: number) {
     ethers.provider.send('evm_mine', []);
 }
 
-describe("Alluo Vault Upgradeable Tests", function () {
+describe("Dola Frax Alluo Vault Upgradeable Tests", function () {
 
     let signers: SignerWithAddress[];
     let usdc: IERC20MetadataUpgradeable, usdt: IERC20MetadataUpgradeable, frax: IERC20MetadataUpgradeable,
-        crv: IERC20MetadataUpgradeable, cvx: IERC20MetadataUpgradeable, rewardToken: IERC20MetadataUpgradeable,
-        yCRVToken: IERC20MetadataUpgradeable, ycrv: IERC20MetadataUpgradeable;
+        crv: IERC20MetadataUpgradeable, cvx: IERC20MetadataUpgradeable, weth: IERC20MetadataUpgradeable,
+        dolaStableCoin: IERC20MetadataUpgradeable, fraxBP: IERC20MetadataUpgradeable,
+        dolaFraxbpToken: IERC20MetadataUpgradeable, rewardToken: IERC20MetadataUpgradeable;
+    let cvxBooster: ICvxBooster;
     let exchange: Exchange;
     const ZERO_ADDR = ethers.constants.AddressZero;
-    let AlluoVault: AlluoVaultUpgradeable;
+    let AlluoVault: AlluoVaultUpgradeable, AlluoVault1: AlluoVaultUpgradeable;
     let alluoPool: IAlluoPool;
-    let yCRVPool: ICurvePool;
+    let dolaFraxMeta: ICurveMetaPool;
     let admin: SignerWithAddress;
 
     async function getImpersonatedSigner(address: string): Promise<SignerWithAddress> {
@@ -50,6 +52,7 @@ describe("Alluo Vault Upgradeable Tests", function () {
     before(async () => {
 
         await resetNetwork();
+
         console.log('\n', "||| Confirm that the _grantRoles(.., msg.sender) in AlluoVaultUpgradeable.sol has been uncommented to ensure tests are functioning correctly |||", '\n')
 
     });
@@ -57,48 +60,57 @@ describe("Alluo Vault Upgradeable Tests", function () {
     beforeEach(async () => {
 
         await resetNetwork();
-
         signers = await ethers.getSigners();
+
         usdc = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48")
         usdt = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xdAC17F958D2ee523a2206206994597C13D831ec7");
         frax = await ethers.getContractAt('IERC20MetadataUpgradeable', '0x853d955acef822db058eb8505911ed77f175b99e');
         crv = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xD533a949740bb3306d119CC777fa900bA034cd52");
         cvx = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x4e3FBD56CD56c3e72c1403e103b45Db9da5B9D2B");
+        weth = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2")
+        cvxBooster = await ethers.getContractAt("ICvxBooster", "0xF403C135812408BFbE8713b5A23a04b3D48AAE31");
         exchange = await ethers.getContractAt("Exchange", "0x29c66CF57a03d41Cfe6d9ecB6883aa0E2AbA21Ec")
+        dolaFraxbpToken = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xE57180685E3348589E9521aa53Af0BCD497E884d");
         rewardToken = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x3A283D9c08E8b55966afb64C515f5143cf907611");
-        yCRVPool = await ethers.getContractAt("ICurvePool", "0x453d92c7d4263201c69aacfaf589ed14202d83a4");
-        yCRVToken = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x453d92c7d4263201c69aacfaf589ed14202d83a4");
-        ycrv = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xfcc5c47be19d06bf83eb04298b026f81069ff65b");
+        dolaStableCoin = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x865377367054516e17014CcdED1e7d814EDC9ce4");
+        fraxBP = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x3175Df0976dFA876431C2E9eE6Bc45b65d3473CC");
+        dolaFraxMeta = await ethers.getContractAt("ICurveMetaPool", "0x08780fb7e580e492c1935bee4fa5920b94aa95da");
         admin = await getImpersonatedSigner("0x1F020A4943EB57cd3b2213A66b355CB662Ea43C3");
 
-        const value = parseEther("10.0");
+        const value = parseEther("2000.0");
 
         await exchange.exchange(
-            ZERO_ADDR, crv.address, value, 0, { value: value }
+            ZERO_ADDR, usdc.address, value, 0, { value: value }
         )
-        const balanceCRV = await crv.balanceOf(signers[0].address);
-        await crv.approve(yCRVPool.address, balanceCRV);
-        await yCRVPool.add_liquidity([balanceCRV, 0], 0);
+
+        await exchange.exchange(
+            ZERO_ADDR, frax.address, value, 0, { value: value }
+        )
+        await exchange.exchange(
+            ZERO_ADDR, usdt.address, value, 0, { value: value }
+        )
+
+        await usdc.approve(dolaFraxMeta.address, value);
+        await dolaFraxMeta["add_liquidity(address,uint256[3],uint256,address)"]("0xe57180685e3348589e9521aa53af0bcd497e884d", [0, 0, 10000], 0, signers[0].address);
 
         let gnosis = "0x1F020A4943EB57cd3b2213A66b355CB662Ea43C3";
         let AlluoVaultFactory = await ethers.getContractFactory("AlluoVaultUpgradeable")
         AlluoVault = await upgrades.deployProxy(AlluoVaultFactory, [
-            "CRV-yCRV Vault",
-            "yCRV",
-            yCRVToken.address, // underlying
+            "Dola-FraxBP Vault",
+            "dolaFrax",
+            dolaFraxbpToken.address, // underlying token
             rewardToken.address, // Curve CVX-ETH Convex Deposit (cvxcrvCVX...)
             ZERO_ADDR, // set pool later
             gnosis,
             "0x84a0856b038eaAd1cC7E297cF34A7e72685A8693", // trusted wallet for meta transactions
-            [crv.address, cvx.address], //
-            [crv.address, ycrv.address], // entry tokens to curve pool
-            124,
-            yCRVPool.address
+            [crv.address, cvx.address], // 
+            [dolaStableCoin.address, fraxBP.address, usdc.address], // entry tokens to curve pool
+            115,
+            dolaFraxbpToken.address
         ], {
             initializer: 'initialize',
             kind: 'uups'
         }) as AlluoVaultUpgradeable;
-
 
         alluoPool = await ethers.getContractAt("AlluoVaultPool", "0x470e486acA0e215C925ddcc3A9D446735AabB714");
         await alluoPool.connect(admin).editVault(true, AlluoVault.address);
@@ -113,7 +125,7 @@ describe("Alluo Vault Upgradeable Tests", function () {
         //     "0xB576491F1E6e5E62f1d8F26062Ee822B40B0E0d4", // Pool address for boosting (CRV/eth)
         //     64, //Pool number convex
         //     cvx.address
-        // ]) as AlluoVaultPool
+        // ]) as AlluoVaultPool;
 
         await AlluoVault.setPool(alluoPool.address);
         await AlluoVault.grantRole("0x0000000000000000000000000000000000000000000000000000000000000000", alluoPool.address)
@@ -124,19 +136,23 @@ describe("Alluo Vault Upgradeable Tests", function () {
         expect(await AlluoVault.totalSupply()).equal(await AlluoVault.totalAssets());
     });
 
-    it("yCRV: Deposit some LP", async function () {
+    it("FRAXBP: Deposit some LP", async function () {
 
-        const lpBalance = await yCRVToken.balanceOf(signers[0].address);
-        await yCRVToken.approve(AlluoVault.address, lpBalance);
+        const lpBalance = await dolaFraxbpToken.balanceOf(signers[0].address);
+        console.log("LP balance before depositing", lpBalance)
+        await dolaFraxbpToken.approve(AlluoVault.address, lpBalance);
         await AlluoVault.deposit(lpBalance, signers[0].address);
+        console.log("Shares after", await AlluoVault.balanceOf(signers[0].address));
+        console.log('LP tokens after', await dolaFraxbpToken.balanceOf(signers[0].address));
         expect(Number(await AlluoVault.balanceOf(signers[0].address))).greaterThan(0);
         expect(Number(lpBalance)).equal(Number(await AlluoVault.balanceOf(signers[0].address)));
 
     })
 
-    it("yCRV: Deposit some LP and wait for rewards to accumulate", async function () {
-        const lpBalance = await yCRVToken.balanceOf(signers[0].address);
-        await yCRVToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
+    it("FRAXBP: Deposit some LP and wait for rewards to accumulate", async function () {
+        const lpBalance = await dolaFraxbpToken.balanceOf(signers[0].address);
+        console.log("LP balance before", lpBalance)
+        await dolaFraxbpToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
         await AlluoVault.stakeUnderlying();
         await skipDays(10);
@@ -165,22 +181,24 @@ describe("Alluo Vault Upgradeable Tests", function () {
         expect(Number(cvxAccumulated)).greaterThan(0)
     });
 
-    it("yCRV: Wait for rewards then loop rewards.", async function () {
-        const lpBalance = await yCRVToken.balanceOf(signers[0].address);
-        await yCRVToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
+    it("FRAXBP: Wait for rewards then loop rewards.", async function () {
+        const lpBalance = await dolaFraxbpToken.balanceOf(signers[0].address);
+        console.log("LP balance before", lpBalance)
+        await dolaFraxbpToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
         await AlluoVault.stakeUnderlying();
         await skipDays(10);
         const fundsBefore = await alluoPool.fundsLocked();
-        console.log("Shareholder accumulated", await AlluoVault.shareholderAccruedRewards(signers[0].address));
         await alluoPool.connect(admin).farm();
 
         console.log("LPs staked", await alluoPool.fundsLocked());
         expect(Number(await alluoPool.fundsLocked())).greaterThan(Number(fundsBefore));
     })
-    it("yCRV: Deposit some Lp for vault tokens and then burn them for the same LPs back.", async function () {
-        const lpBalance = await yCRVToken.balanceOf(signers[0].address);
-        await yCRVToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
+
+    it("FRAXBP: Deposit some Lp for vault tokens and then burn them for the same LPs back.", async function () {
+        const lpBalance = await dolaFraxbpToken.balanceOf(signers[0].address);
+        console.log('balance', lpBalance);
+        await dolaFraxbpToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
 
         await AlluoVault.stakeUnderlying();
@@ -194,26 +212,27 @@ describe("Alluo Vault Upgradeable Tests", function () {
 
         expect(await AlluoVault.balanceOf(signers[0].address)).equal(0)
         // this must be greater than or equal because the base pool APY is quite high.
-        expect(Number(await yCRVToken.balanceOf(signers[0].address))).greaterThanOrEqual(Number(lpBalance))
+        console.log('LP tokens after withdrawal', await dolaFraxbpToken.balanceOf(signers[0].address));
+        expect(Number(await dolaFraxbpToken.balanceOf(signers[0].address))).greaterThanOrEqual(Number(lpBalance))
         console.log("Rewards in LP", Number(await rewardToken.balanceOf(signers[0].address)));
         expect(Number(await rewardToken.balanceOf(signers[0].address))).greaterThan(0)
         expect(Number(await AlluoVault.totalAssets())).equal(0)
         expect(Number(await AlluoVault.totalSupply())).equal(0)
     })
 
-    it("yCRV: After looping rewards, expect fundsLocked to increase.", async function () {
+    it("FRAXBP: After looping rewards, expect fundsLocked to increase.", async function () {
 
-        const lpBalance = await yCRVToken.balanceOf(signers[0].address);
-        await yCRVToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
+        const lpBalance = await dolaFraxbpToken.balanceOf(signers[0].address);
+        await dolaFraxbpToken.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
         await AlluoVault.stakeUnderlying();
-        console.log('\nfunds locked', await alluoPool.fundsLocked());
         await skipDays(0.1);
         await alluoPool.connect(admin).farm();
 
         const initialRewards = await alluoPool.fundsLocked();
+        console.log('LPs staked before', initialRewards);
 
-        await skipDays(0.5);
+        await skipDays(10);
         await alluoPool.claimRewardsFromPool();
 
         const crvAccumulated = await crv.balanceOf(alluoPool.address);
@@ -228,16 +247,14 @@ describe("Alluo Vault Upgradeable Tests", function () {
         expect(Number(compoundedRewards)).greaterThan(Number(initialRewards));
     })
 
-    it("yCRV: Deposit frax to enter pool.", async function () {
+    it("FRAXBP: Deposit frax to enter pool.", async function () {
 
-        await exchange.exchange(
-            ZERO_ADDR, frax.address, parseEther("100"), 0, { value: parseEther("100") }
-        )
-        const fraxBalance = parseEther("10");
+        const fraxBalance = parseEther("100");
         await frax.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.depositWithoutLP(fraxBalance, frax.address);
         await AlluoVault.stakeUnderlying();
-        await skipDays(5);
+        await skipDays(10);
+        console.log("Shareholder accumulated", await AlluoVault.shareholderAccruedRewards(signers[0].address));
 
         await AlluoVault.claimRewardsFromPool();
 
@@ -247,17 +264,15 @@ describe("Alluo Vault Upgradeable Tests", function () {
         expect(Number(cvxAccumulated)).greaterThan(0)
     })
 
-    it("yCRV: Deposit usdc to enter pool.", async function () {
+    it("FRAXBP: Deposit usdc to enter pool.", async function () {
         const usdcBalance = await usdc.balanceOf(signers[0].address);
-        await exchange.exchange(
-            ZERO_ADDR, usdc.address, parseEther("100"), 0, { value: parseEther("100") }
-        )
         await usdc.approve(AlluoVault.address, usdcBalance);
 
         const crvBefore = await crv.balanceOf(AlluoVault.address);
         const cvxBefore = await cvx.balanceOf(AlluoVault.address);
 
         await AlluoVault.depositWithoutLP(usdcBalance, usdc.address);
+        expect(usdcBalance).to.be.gt(await usdc.balanceOf(signers[0].address));
         await AlluoVault.stakeUnderlying();
         await skipDays(10);
         await AlluoVault.claimRewardsFromPool();
@@ -268,16 +283,14 @@ describe("Alluo Vault Upgradeable Tests", function () {
         expect(Number(cvxAccumulated)).greaterThan(Number(cvxBefore));
     })
 
-    it("yCRV: Deposit usdt to enter pool (non pool token).", async function () {
-        await exchange.exchange(
-            ZERO_ADDR, usdt.address, parseEther("100"), 0, { value: parseEther("100") }
-        )
+    it("FRAXBP: Deposit usdt to enter pool (non pool token).", async function () {
+
         const usdtBalance = parseUnits("100", 6);
         const crvBefore = await crv.balanceOf(AlluoVault.address);
         const cvxBefore = await cvx.balanceOf(AlluoVault.address);
-
         await usdt.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.depositWithoutLP(usdtBalance, usdt.address);
+        expect(await usdt.balanceOf(signers[0].address)).to.be.gt(usdtBalance);
         await AlluoVault.stakeUnderlying();
         await skipDays(10);
         await AlluoVault.claimRewardsFromPool();
@@ -288,28 +301,23 @@ describe("Alluo Vault Upgradeable Tests", function () {
         expect(Number(cvxAccumulated)).greaterThan(Number(cvxBefore));
     })
 
-    it("yCRV: Deposit usdc to enter pool and exit again in USDC", async function () {
+    it("FRAXBP: Deposit usdc to enter pool and exit again in USDC", async function () {
 
-        await exchange.exchange(
-            ZERO_ADDR, usdc.address, parseEther("100"), 0, { value: parseEther("100") }
-        )
         const usdcBalance = parseUnits("100", 6);
         await usdc.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.depositWithoutLP(usdcBalance, usdc.address);
+        expect(await usdc.balanceOf(signers[0].address)).to.be.gt(usdcBalance);
         await AlluoVault.stakeUnderlying();
         await skipDays(10);
         await AlluoVault.claimRewardsFromPool();
 
         const lpBalance = await AlluoVault.balanceOf(signers[0].address);
-        await AlluoVault.withdrawToNonLp(lpBalance, signers[0].address, signers[0].address, usdc.address);
+        await AlluoVault.withdrawToNonLp(lpBalance, signers[0].address, signers[0].address, usdc.address)
         expect(await AlluoVault.balanceOf(signers[0].address)).to.be.eq(0);
 
     })
 
-    it("yCRV: Deposit usdt to enter pool (non pool token) and exit again in a non pool token.", async function () {
-        await exchange.exchange(
-            ZERO_ADDR, usdt.address, parseEther("100"), 0, { value: parseEther("100") }
-        )
+    it("FRAXBP: Deposit usdt to enter pool (non pool token) and exit again in a non pool token.", async function () {
         const usdtBalance = parseUnits("100", 6);
         await usdt.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.depositWithoutLP(usdtBalance, usdt.address);
@@ -318,19 +326,19 @@ describe("Alluo Vault Upgradeable Tests", function () {
         await skipDays(10);
         await AlluoVault.claimRewardsFromPool();
         const lpBalance = await AlluoVault.balanceOf(signers[0].address);
-        await AlluoVault.withdrawToNonLp(lpBalance, signers[0].address, signers[0].address, usdt.address);
+        await AlluoVault.withdrawToNonLp(lpBalance, signers[0].address, signers[0].address, usdt.address)
         expect(await usdt.balanceOf(signers[0].address)).to.be.gt(usdtBefore);
-        console.log(usdtBefore, await usdt.balanceOf(signers[0].address));
+
     })
 
-    it("yCRV: Multiple deposits and withdrawals should return correct LP amounts", async function () {
+    it("FRAXBP: Multiple deposits and withdrawals should return correct LP amounts", async function () {
         let signerBalancesBefore = []
         for (let i = 1; i < 6; i++) {
             await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, yCRVToken.address, parseEther("10"), 0, { value: parseEther("10") }
+                ZERO_ADDR, dolaFraxbpToken.address, parseEther("10"), 0, { value: parseEther("10") }
             )
-            let lpBalance = await yCRVToken.balanceOf(signers[i].address)
-            await yCRVToken.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
+            let lpBalance = await dolaFraxbpToken.balanceOf(signers[i].address)
+            await dolaFraxbpToken.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
             await AlluoVault.connect(signers[i]).deposit(lpBalance, signers[i].address);
             console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
             signerBalancesBefore.push(await AlluoVault.balanceOf(signers[i].address));
@@ -345,16 +353,16 @@ describe("Alluo Vault Upgradeable Tests", function () {
         }
 
         for (let i = 0; i < 5; i++) {
-            expect(Number(signerBalancesBefore[i])).equal(Number(await yCRVToken.balanceOf(signers[i + 1].address)))
+            expect(Number(signerBalancesBefore[i])).equal(Number(await dolaFraxbpToken.balanceOf(signers[i + 1].address)))
         }
     })
-    it("yCRV: Multiple deposits should return correct LP amounts and reward distribution (equal here)", async function () {
+    it("FRAXBP: Multiple deposits should return correct LP amounts and reward distribution (equal here)", async function () {
         for (let i = 1; i < 6; i++) {
             await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, yCRVToken.address, parseEther("10"), 0, { value: parseEther("10") }
+                ZERO_ADDR, dolaFraxbpToken.address, parseEther("10"), 0, { value: parseEther("10") }
             )
             let lpBalance = parseEther("100")
-            await yCRVToken.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
+            await dolaFraxbpToken.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
             await AlluoVault.connect(signers[i]).deposit(lpBalance, signers[i].address);
             console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
         }
@@ -376,7 +384,7 @@ describe("Alluo Vault Upgradeable Tests", function () {
     })
 
 
-    it("yCRV: Multiple deposits and withdrawals in nonLP tokens should return correct LP amounts and reward distribution (equal here)", async function () {
+    it("FRAXBP: Multiple deposits and withdrawals in nonLP tokens should return correct LP amounts and reward distribution (equal here)", async function () {
         for (let i = 1; i < 6; i++) {
             await exchange.connect(signers[i]).exchange(
                 ZERO_ADDR, frax.address, parseEther("10"), 0, { value: parseEther("10") }
@@ -410,14 +418,14 @@ describe("Alluo Vault Upgradeable Tests", function () {
 
     })
 
-    it("yCRV: After some loops, the multisig should be able to claim fees accumulated.", async function () {
+    it("FRAXBP: After some loops, the multisig should be able to claim fees accumulated.", async function () {
         await AlluoVault.setAdminFee(100);
         for (let i = 1; i < 6; i++) {
             await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, yCRVToken.address, parseEther("10"), 0, { value: parseEther("10") }
+                ZERO_ADDR, dolaFraxbpToken.address, parseEther("10"), 0, { value: parseEther("10") }
             )
-            let lpBalance = await yCRVToken.balanceOf(signers[i].address)
-            await yCRVToken.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
+            let lpBalance = await dolaFraxbpToken.balanceOf(signers[i].address)
+            await dolaFraxbpToken.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
             await AlluoVault.connect(signers[i]).deposit(lpBalance, signers[i].address);
             console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
         }
