+++ test/boost/AlluoVaultUpgradeableStEth.ts
@@ -5,7 +5,7 @@ import { BigNumber, BigNumberish, BytesLike, Wallet } from "ethers";
 import { defaultAbiCoder } from "ethers/lib/utils";
 import { ethers, network, upgrades } from "hardhat";
 import { afterEach, before } from "mocha";
-import { AlluoVaultUpgradeable, Exchange, AlluoVaultPool, IAlluoPool, ICurvePool, ICvxBooster, IERC20MetadataUpgradeable, IExchange } from "../typechain";
+import { AlluoVaultUpgradeable, Exchange, AlluoVaultPool, IAlluoPool, ICurvePool, ICvxBooster, IERC20MetadataUpgradeable, IExchange, IWrappedEther } from "../../typechain";
 
 
 async function getImpersonatedSigner(address: string): Promise<SignerWithAddress> {
@@ -22,22 +22,39 @@ async function skipDays(d: number) {
     ethers.provider.send('evm_mine', []);
 }
 
-describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
+describe("StEth Alluo Vault Upgradeable Tests", function () {
     let signers: SignerWithAddress[];
-    let usdc: IERC20MetadataUpgradeable, usdt: IERC20MetadataUpgradeable, frax: IERC20MetadataUpgradeable, crv: IERC20MetadataUpgradeable, cvx: IERC20MetadataUpgradeable, weth: IERC20MetadataUpgradeable;
+    let usdc: IERC20MetadataUpgradeable, usdt: IERC20MetadataUpgradeable, frax: IERC20MetadataUpgradeable, crv: IERC20MetadataUpgradeable, cvx: IERC20MetadataUpgradeable, weth: IERC20MetadataUpgradeable, ldo: IERC20MetadataUpgradeable, stEth: IERC20MetadataUpgradeable;
     let cvxBooster: ICvxBooster;
     let exchange: Exchange;
     const ZERO_ADDR = ethers.constants.AddressZero;
     let AlluoVault: AlluoVaultUpgradeable;
     let rewardToken: IERC20MetadataUpgradeable;
-    let fraxUSDC: IERC20MetadataUpgradeable;
+    let stEthEth: IERC20MetadataUpgradeable;
     let alluoPool: IAlluoPool;
+
+    before(async function () {
+        await network.provider.request({
+            method: "hardhat_reset",
+            params: [{
+                forking: {
+                    enabled: true,
+                    jsonRpcUrl: process.env.MAINNET_FORKING_URL as string,
+                    //you can fork from last block by commenting next line
+                    blockNumber: 15506073,
+                },
+            },],
+        });
+
+    })
 
     before(async () => {
         console.log('\n', "||| Confirm that the _grantRoles(.., msg.sender) in AlluoVaultUpgradeable.sol has been uncommented to ensure tests are functioning correctly |||", '\n')
-    })
+
+    });
 
     beforeEach(async () => {
+
         await network.provider.request({
             method: "hardhat_reset",
             params: [{
@@ -56,14 +73,18 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
         frax = await ethers.getContractAt('IERC20MetadataUpgradeable', '0x853d955acef822db058eb8505911ed77f175b99e');
         crv = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xD533a949740bb3306d119CC777fa900bA034cd52");
         cvx = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x4e3FBD56CD56c3e72c1403e103b45Db9da5B9D2B");
+        ldo = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x5A98FcBEA516Cf06857215779Fd812CA3beF1B32");
+
+        stEth = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xae7ab96520DE3A18E5e111B5EaAb095312D7fE84");
         weth = await ethers.getContractAt("IERC20MetadataUpgradeable", "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2")
         cvxBooster = await ethers.getContractAt("ICvxBooster", "0xF403C135812408BFbE8713b5A23a04b3D48AAE31");
         exchange = await ethers.getContractAt("Exchange", "0x29c66CF57a03d41Cfe6d9ecB6883aa0E2AbA21Ec")
-        fraxUSDC = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x3175Df0976dFA876431C2E9eE6Bc45b65d3473CC");
+        stEthEth = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x06325440D014e39736583c165C2963BA99fAf14E");
         rewardToken = await ethers.getContractAt("IERC20MetadataUpgradeable", "0x3A283D9c08E8b55966afb64C515f5143cf907611");
 
         const value = parseEther("2000.0");
-
+        const wrappedEther = await ethers.getContractAt("contracts/boost/interfaces/IWrappedEther.sol:IWrappedEther", "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2") as IWrappedEther;
+        await wrappedEther.deposit({ value: "100" })
         await exchange.exchange(
             ZERO_ADDR, frax.address, value, 0, { value: value }
         )
@@ -76,31 +97,29 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
         )
 
 
-        const fraxUSDCPool = await ethers.getContractAt("ICurvePool", "0xDcEF968d416a41Cdac0ED8702fAC8128A64241A2")
-        await frax.approve(fraxUSDCPool.address, ethers.constants.MaxUint256)
-        await fraxUSDCPool.add_liquidity([parseEther("1000"), 0], 0);
+        const stEthEthPool = await ethers.getContractAt("ICurvePool", "0xDC24316b9AE028F1497c275EB9192a3Ea0f67022")
+        await stEthEthPool.add_liquidity([parseEther("1"), 0], 0, { value: parseEther("1") });
         let gnosis = "0x6b140e772aCC4D5E0d5Eac3813D586aa6DB8Fbf7";
         let AlluoVaultFactory = await ethers.getContractFactory("AlluoVaultUpgradeable")
 
         AlluoVault = await upgrades.deployProxy(AlluoVaultFactory, [
-            "Frax-USDC Vault",
-            "abFraxUSDC",
-            fraxUSDC.address,
+            "StEth-Eth Vault",
+            "abStETH",
+            stEthEth.address,
             rewardToken.address,
             rewardToken.address,
             gnosis,
             "0x84a0856b038eaAd1cC7E297cF34A7e72685A8693",
-            [crv.address, cvx.address],
-            [frax.address, usdc.address],
-            100,
-            fraxUSDCPool.address
+            [crv.address, cvx.address, ldo.address],
+            ["0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE", stEth.address],
+            25,
+            stEthEthPool.address
         ], {
             initializer: 'initialize',
             kind: 'uups'
         }) as AlluoVaultUpgradeable;
 
         let PoolVaultFactory = await ethers.getContractFactory("AlluoVaultPool");
-
         alluoPool = await upgrades.deployProxy(PoolVaultFactory, [
             rewardToken.address,
             gnosis,
@@ -113,16 +132,15 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
         await AlluoVault.setPool(alluoPool.address);
 
         await AlluoVault.grantRole("0x0000000000000000000000000000000000000000000000000000000000000000", alluoPool.address)
-
     });
 
     afterEach(async () => {
         expect(await AlluoVault.totalSupply()).equal(await AlluoVault.totalAssets());
     })
     it("Deposit some LP", async function () {
-        const lpBalance = await fraxUSDC.balanceOf(signers[0].address);
-        console.log("Balance before of FRAX-USDC Lp", lpBalance)
-        await fraxUSDC.approve(AlluoVault.address, ethers.constants.MaxUint256);
+        const lpBalance = await stEthEth.balanceOf(signers[0].address);
+        console.log("Balance before of stEth-Eth Lp", lpBalance)
+        await stEthEth.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
         console.log("Shares after", await AlluoVault.balanceOf(signers[0].address));
         expect(Number(await AlluoVault.balanceOf(signers[0].address))).greaterThan(0);
@@ -130,9 +148,9 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
     })
 
     it("Deposit some LP and wait for rewards to accumulate", async function () {
-        const lpBalance = await fraxUSDC.balanceOf(signers[0].address);
-        console.log("Balance before of FRAX-USDC Lp", lpBalance)
-        await fraxUSDC.approve(AlluoVault.address, ethers.constants.MaxUint256);
+        const lpBalance = await stEthEth.balanceOf(signers[0].address);
+        console.log("Balance before of stEth-Eth Lp", lpBalance)
+        await stEthEth.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
         await AlluoVault.stakeUnderlying();
         await skipDays(0.01);
@@ -140,15 +158,21 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
 
         const crvAccumulated = await crv.balanceOf(AlluoVault.address);
         const cvxAccumulated = await cvx.balanceOf(AlluoVault.address);
+        const ldoAccumulated = await ldo.balanceOf(AlluoVault.address);
+
         console.log(crvAccumulated)
         console.log(cvxAccumulated)
+        console.log(ldoAccumulated)
+
         expect(Number(crvAccumulated)).greaterThan(0)
         expect(Number(cvxAccumulated)).greaterThan(0)
+        expect(Number(ldoAccumulated)).greaterThan(0)
+
     })
     it("Wait for rewards then loop rewards.", async function () {
-        const lpBalance = await fraxUSDC.balanceOf(signers[0].address);
-        console.log("Balance before of FRAX-USDC Lp", lpBalance)
-        await fraxUSDC.approve(AlluoVault.address, ethers.constants.MaxUint256);
+        const lpBalance = await stEthEth.balanceOf(signers[0].address);
+        console.log("Balance before of stEth-Eth Lp", lpBalance)
+        await stEthEth.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
         await AlluoVault.stakeUnderlying();
         await skipDays(0.01);
@@ -158,29 +182,11 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
         console.log("crv-ETH staked", await alluoPool.fundsLocked());
         expect(Number(await alluoPool.fundsLocked())).greaterThan(0);
     })
-    it("Deposit some Lp for vault tokens and then burn them for the same LPs back.", async function () {
-        const lpBalance = await fraxUSDC.balanceOf(signers[0].address);
-        await fraxUSDC.approve(AlluoVault.address, ethers.constants.MaxUint256);
-        await AlluoVault.deposit(lpBalance, signers[0].address);
-        await AlluoVault.stakeUnderlying();
-        await skipDays(0.01);
-        await AlluoVault.claimRewardsFromPool();
-        await alluoPool.farm();
 
-        await AlluoVault.withdraw(lpBalance, signers[0].address, signers[0].address);
-        await AlluoVault.claimRewards();
-
-        expect(await AlluoVault.balanceOf(signers[0].address)).equal(0)
-        expect(await fraxUSDC.balanceOf(signers[0].address)).equal(lpBalance)
-        console.log("Rewardsin LP", Number(await rewardToken.balanceOf(signers[0].address)));
-        expect(Number(await rewardToken.balanceOf(signers[0].address))).greaterThan(0)
-        expect(Number(await AlluoVault.totalAssets())).equal(0)
-        expect(Number(await AlluoVault.totalSupply())).equal(0)
-    })
     it("After looping rewards, expect fundsLocked to increase.", async function () {
-        const lpBalance = await fraxUSDC.balanceOf(signers[0].address);
-        console.log("Balance before of FRAX-USDC Lp", lpBalance)
-        await fraxUSDC.approve(AlluoVault.address, ethers.constants.MaxUint256);
+        const lpBalance = await stEthEth.balanceOf(signers[0].address);
+        console.log("Balance before of stEth-Eth Lp", lpBalance)
+        await stEthEth.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.deposit(lpBalance, signers[0].address);
         await AlluoVault.stakeUnderlying();
         await skipDays(0.01);
@@ -195,8 +201,11 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
 
         const crvAccumulated = await crv.balanceOf(alluoPool.address);
         const cvxAccumulated = await cvx.balanceOf(alluoPool.address);
+        const ldoAccumulated = await ldo.balanceOf(AlluoVault.address);
+
         console.log(crvAccumulated)
         console.log(cvxAccumulated)
+        console.log(ldoAccumulated)
 
 
         await alluoPool.farm();
@@ -205,9 +214,28 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
         console.log("crv-ETH staked after", await alluoPool.fundsLocked());
         expect(Number(compoundedRewards)).greaterThan(Number(initialRewards));
     })
+
+    it("Deposit some Lp for vault tokens and then burn them for the same LPs back.", async function () {
+        const lpBalance = await stEthEth.balanceOf(signers[0].address);
+        await stEthEth.approve(AlluoVault.address, ethers.constants.MaxUint256);
+        await AlluoVault.deposit(lpBalance, signers[0].address);
+        await AlluoVault.stakeUnderlying();
+        await skipDays(0.01);
+        await AlluoVault.claimRewardsFromPool();
+        await alluoPool.farm();
+
+        await AlluoVault.withdraw(lpBalance, signers[0].address, signers[0].address);
+        await AlluoVault.claimRewards();
+
+        expect(await AlluoVault.balanceOf(signers[0].address)).equal(0)
+        expect(await stEthEth.balanceOf(signers[0].address)).equal(lpBalance)
+        console.log("Rewardsin LP", Number(await rewardToken.balanceOf(signers[0].address)));
+        expect(Number(await rewardToken.balanceOf(signers[0].address))).greaterThan(0)
+        expect(Number(await AlluoVault.totalAssets())).equal(0)
+        expect(Number(await AlluoVault.totalSupply())).equal(0)
+    })
+
     it("Deposit frax to enter pool.", async function () {
-        // const fraxBalance = await frax.balanceOf(signers[0].address);
-        // console.log("Balance before of Frax balance", fraxBalance)
         const fraxBalance = parseEther("100");
         await frax.approve(AlluoVault.address, ethers.constants.MaxUint256);
         await AlluoVault.depositWithoutLP(fraxBalance, frax.address);
@@ -217,28 +245,36 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
 
         const crvAccumulated = await crv.balanceOf(AlluoVault.address);
         const cvxAccumulated = await cvx.balanceOf(AlluoVault.address);
+        const ldoAccumulated = await ldo.balanceOf(AlluoVault.address);
+
         console.log(crvAccumulated)
         console.log(cvxAccumulated)
+        console.log(ldoAccumulated)
+
         expect(Number(crvAccumulated)).greaterThan(0)
         expect(Number(cvxAccumulated)).greaterThan(0)
+        expect(Number(ldoAccumulated)).greaterThan(0)
     })
 
-    it("Deposit usdc to enter pool.", async function () {
-        // const fraxBalance = await frax.balanceOf(signers[0].address);
-        // console.log("Balance before of Frax balance", fraxBalance)
-        const usdcBalance = parseUnits("100", 6);
+    it("Deposit StEth to enter pool.", async function () {
+        const stEth = parseUnits("100", 6);
         await usdc.approve(AlluoVault.address, ethers.constants.MaxUint256);
-        await AlluoVault.depositWithoutLP(usdcBalance, usdc.address);
+        await AlluoVault.depositWithoutLP(stEth, usdc.address);
         await AlluoVault.stakeUnderlying();
         await skipDays(0.01);
         await AlluoVault.claimRewardsFromPool();
 
         const crvAccumulated = await crv.balanceOf(AlluoVault.address);
         const cvxAccumulated = await cvx.balanceOf(AlluoVault.address);
+        const ldoAccumulated = await ldo.balanceOf(AlluoVault.address);
+
         console.log(crvAccumulated)
         console.log(cvxAccumulated)
+        console.log(ldoAccumulated)
+
         expect(Number(crvAccumulated)).greaterThan(0)
         expect(Number(cvxAccumulated)).greaterThan(0)
+        expect(Number(ldoAccumulated)).greaterThan(0)
     })
 
     it("Deposit usdt to enter pool (non pool token).", async function () {
@@ -252,10 +288,15 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
 
         const crvAccumulated = await crv.balanceOf(AlluoVault.address);
         const cvxAccumulated = await cvx.balanceOf(AlluoVault.address);
+        const ldoAccumulated = await ldo.balanceOf(AlluoVault.address);
+
         console.log(crvAccumulated)
         console.log(cvxAccumulated)
+        console.log(ldoAccumulated)
+
         expect(Number(crvAccumulated)).greaterThan(0)
         expect(Number(cvxAccumulated)).greaterThan(0)
+        expect(Number(ldoAccumulated)).greaterThan(0)
     })
 
     it("Deposit usdc to enter pool and exit again in USDC", async function () {
@@ -287,14 +328,13 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
     it("Multiple deposits and withdrawals should return correct LP amounts", async function () {
         let signerBalancesBefore = []
         for (let i = 1; i < 6; i++) {
-            await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, frax.address, parseEther("10"), 0, { value: parseEther("10") }
-            )
-            const fraxUSDCPool = await ethers.getContractAt("ICurvePool", "0xDcEF968d416a41Cdac0ED8702fAC8128A64241A2")
-            await frax.connect(signers[i]).approve(fraxUSDCPool.address, ethers.constants.MaxUint256)
-            await fraxUSDCPool.connect(signers[i]).add_liquidity([parseEther("1000"), 0], 0);
-            let lpBalance = await fraxUSDC.balanceOf(signers[i].address)
-            await fraxUSDC.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
+            const wrappedEther = await ethers.getContractAt("contracts/boost/interfaces/IWrappedEther.sol:IWrappedEther", "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2") as IWrappedEther;
+            await wrappedEther.connect(signers[i]).deposit({ value: "1" })
+            const stEthEthPool = await ethers.getContractAt("ICurvePool", "0xDC24316b9AE028F1497c275EB9192a3Ea0f67022")
+            await stEthEthPool.connect(signers[i]).add_liquidity([parseEther("1"), 0], 0, { value: parseEther("1") });
+
+            let lpBalance = await stEthEth.balanceOf(signers[i].address)
+            await stEthEth.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
             await AlluoVault.connect(signers[i]).deposit(lpBalance, signers[i].address);
             console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
             signerBalancesBefore.push(await AlluoVault.balanceOf(signers[i].address));
@@ -309,105 +349,36 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
         }
 
         for (let i = 0; i < 5; i++) {
-            expect(Number(signerBalancesBefore[i])).equal(Number(await fraxUSDC.balanceOf(signers[i + 1].address)))
+            expect(Number(signerBalancesBefore[i])).equal(Number(await stEthEth.balanceOf(signers[i + 1].address)))
         }
     })
-    it("Multiple deposits should return correct LP amounts and reward distribution (equal here)", async function () {
-        for (let i = 1; i < 6; i++) {
-            await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, frax.address, parseEther("10"), 0, { value: parseEther("10") }
-            )
-            const fraxUSDCPool = await ethers.getContractAt("ICurvePool", "0xDcEF968d416a41Cdac0ED8702fAC8128A64241A2")
-            await frax.connect(signers[i]).approve(fraxUSDCPool.address, ethers.constants.MaxUint256)
-            await fraxUSDCPool.connect(signers[i]).add_liquidity([parseEther("1000"), 0], 0);
-            let lpBalance = await fraxUSDC.balanceOf(signers[i].address)
-            await fraxUSDC.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
-            await AlluoVault.connect(signers[i]).deposit(lpBalance, signers[i].address);
-            console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
-        }
-        await AlluoVault.stakeUnderlying();
-        await skipDays(0.01);
-        await AlluoVault.claimRewardsFromPool();
-        await alluoPool.farm();
-        await skipDays(0.01);
-        await AlluoVault.connect(signers[1]).claimRewards();
-        let expectBalance = await rewardToken.balanceOf(signers[1].address)
-
-        for (let i = 2; i < 6; i++) {
-            await AlluoVault.connect(signers[i]).claimRewards();
-            // Small dust
-            expect(Number(await rewardToken.balanceOf(signers[i].address)).toPrecision(5)).equal(Number(expectBalance).toPrecision(5))
-            console.log(`Reward tokens for signer ${i}: ${await rewardToken.balanceOf(signers[i].address)}`)
-        }
-    })
-
-    it("Multiple deposits in nonLP tokens should return correct LP amounts and reward distribution (equal here)", async function () {
-        for (let i = 1; i < 6; i++) {
-            await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, frax.address, parseEther("10"), 0, { value: parseEther("10") }
-            )
-            const fraxBalance = parseEther("100");
-            await frax.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
-            await AlluoVault.connect(signers[i]).depositWithoutLP(fraxBalance, frax.address);
-
-            console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
-        }
-        await AlluoVault.stakeUnderlying();
-        await skipDays(0.01);
-        expect(await AlluoVault.totalSupply()).equal(await AlluoVault.totalAssets());
-
-        await AlluoVault.claimRewardsFromPool();
-        await alluoPool.farm();
-        await skipDays(0.01);
-        await AlluoVault.connect(signers[1]).claimRewards();
-        let expectBalance = await rewardToken.balanceOf(signers[1].address)
-
-        for (let i = 2; i < 6; i++) {
-            await AlluoVault.connect(signers[i]).claimRewards();
-            // Small dust
-            expect(Number(await rewardToken.balanceOf(signers[i].address)).toPrecision(5)).equal(Number(expectBalance).toPrecision(5))
-            console.log(`Reward tokens for signer ${i}: ${await rewardToken.balanceOf(signers[i].address)}`)
-        }
-    })
-
     it("Multiple deposits and withdrawals should return correct LP amounts and reward distribution (equal here)", async function () {
         for (let i = 1; i < 6; i++) {
-            await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, frax.address, parseEther("10"), 0, { value: parseEther("10") }
-            )
-            const fraxUSDCPool = await ethers.getContractAt("ICurvePool", "0xDcEF968d416a41Cdac0ED8702fAC8128A64241A2")
-            await frax.connect(signers[i]).approve(fraxUSDCPool.address, ethers.constants.MaxUint256)
-            await fraxUSDCPool.connect(signers[i]).add_liquidity([parseEther("1000"), 0], 0);
-            let lpBalance = await fraxUSDC.balanceOf(signers[i].address)
-            console.log("lpBalance", lpBalance)
-            await fraxUSDC.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
+            const wrappedEther = await ethers.getContractAt("contracts/boost/interfaces/IWrappedEther.sol:IWrappedEther", "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2") as IWrappedEther;
+            await wrappedEther.connect(signers[i]).deposit({ value: "1" })
+            const stEthEthPool = await ethers.getContractAt("ICurvePool", "0xDC24316b9AE028F1497c275EB9192a3Ea0f67022")
+            await stEthEthPool.connect(signers[i]).add_liquidity([parseEther("1"), 0], 0, { value: parseEther("1") });
+
+            let lpBalance = parseEther("0.5")
+            await stEthEth.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
             await AlluoVault.connect(signers[i]).deposit(lpBalance, signers[i].address);
             console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
         }
+
         await AlluoVault.stakeUnderlying();
         await skipDays(0.01);
         await AlluoVault.claimRewardsFromPool();
         await alluoPool.farm();
         await skipDays(0.01);
-
-        expect(await AlluoVault.totalSupply()).equal(await AlluoVault.totalAssets());
-
         await AlluoVault.connect(signers[1]).claimRewards();
-        await AlluoVault.connect(signers[1]).withdraw(await AlluoVault.balanceOf(signers[1].address), signers[1].address, signers[1].address)
         let expectBalance = await rewardToken.balanceOf(signers[1].address)
-        expect(await AlluoVault.balanceOf(signers[1].address)).equal(0);
 
         for (let i = 2; i < 6; i++) {
             await AlluoVault.connect(signers[i]).claimRewards();
-            await AlluoVault.connect(signers[i]).withdraw(await AlluoVault.balanceOf(signers[i].address), signers[i].address, signers[i].address)
             // Small dust
             expect(Number(await rewardToken.balanceOf(signers[i].address)).toPrecision(5)).equal(Number(expectBalance).toPrecision(5))
-            expect(await AlluoVault.balanceOf(signers[i].address)).equal(0);
-
             console.log(`Reward tokens for signer ${i}: ${await rewardToken.balanceOf(signers[i].address)}`)
         }
-        expect(await AlluoVault.totalSupply()).equal(await AlluoVault.totalAssets());
-
     })
 
     it("Multiple deposits and withdrawals in nonLP tokens should return correct LP amounts and reward distribution (equal here)", async function () {
@@ -435,7 +406,7 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
             await AlluoVault.connect(signers[i]).claimRewards();
             await AlluoVault.connect(signers[i]).withdrawToNonLp(await AlluoVault.balanceOf(signers[i].address), signers[i].address, signers[i].address, frax.address);
             // Small dust
-            expect(Number(await rewardToken.balanceOf(signers[i].address)).toPrecision(3)).equal(Number(expectBalance).toPrecision(3))
+            expect(Number(await rewardToken.balanceOf(signers[i].address)).toPrecision(2)).equal(Number(expectBalance).toPrecision(2))
             expect(await AlluoVault.balanceOf(signers[i].address)).equal(0);
             console.log(`Reward tokens for signer ${i}: ${await rewardToken.balanceOf(signers[i].address)}`)
         }
@@ -443,17 +414,21 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
 
     })
 
+
+
     it("After some loops, the multisig should be able to claim fees accumulated.", async function () {
         await AlluoVault.setAdminFee(100)
+
+        let gnosis = "0x6b140e772aCC4D5E0d5Eac3813D586aa6DB8Fbf7";
+
         for (let i = 1; i < 6; i++) {
-            await exchange.connect(signers[i]).exchange(
-                ZERO_ADDR, frax.address, parseEther("10"), 0, { value: parseEther("10") }
-            )
-            const fraxUSDCPool = await ethers.getContractAt("ICurvePool", "0xDcEF968d416a41Cdac0ED8702fAC8128A64241A2")
-            await frax.connect(signers[i]).approve(fraxUSDCPool.address, ethers.constants.MaxUint256)
-            await fraxUSDCPool.connect(signers[i]).add_liquidity([parseEther("1000"), 0], 0);
-            let lpBalance = await fraxUSDC.balanceOf(signers[i].address)
-            await fraxUSDC.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
+            const wrappedEther = await ethers.getContractAt("contracts/boost/interfaces/IWrappedEther.sol:IWrappedEther", "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2") as IWrappedEther;
+            await wrappedEther.connect(signers[i]).deposit({ value: "1" })
+            const stEthEthPool = await ethers.getContractAt("ICurvePool", "0xDC24316b9AE028F1497c275EB9192a3Ea0f67022")
+            await stEthEthPool.connect(signers[i]).add_liquidity([parseEther("1"), 0], 0, { value: parseEther("1") });
+
+            let lpBalance = await stEthEth.balanceOf(signers[i].address)
+            await stEthEth.connect(signers[i]).approve(AlluoVault.address, ethers.constants.MaxUint256);
             await AlluoVault.connect(signers[i]).deposit(lpBalance, signers[i].address);
             console.log(`Signer ${i}:`, await AlluoVault.balanceOf(signers[i].address));
         }
@@ -463,7 +438,7 @@ describe("Frax USDC Alluo Vault Upgradeable Tests", function () {
         await alluoPool.farm();
         await skipDays(0.01);
 
-        let gnosis = "0x6b140e772aCC4D5E0d5Eac3813D586aa6DB8Fbf7";
         expect(Number(await AlluoVault.earned(gnosis))).greaterThan(0);
+
     })
 })
